refactor: rename http_routes module to http (#751)
